Data type change

+++ include/SensorFusionAlgorithm.h
@@ -6,8 +6,8 @@
 #include <stdio.h>
 
 double** SDMCalculator(int); // This function will calculate the support degree matrix and return it to the requesting function
-int* PrincipalComponents(); //This function will calculate all the principal components of the D matrix
-int* ComputationScore(); //This function will calculate and eliminate the incorrect data
-int* OutputCompute(); // This function will return the fused output to the main function
+double* PrincipalComponents(); //This function will calculate all the principal components of the D matrix
+double* ComputationScore(); //This function will calculate and eliminate the incorrect data
+double* OutputCompute(); // This function will return the fused output to the main function
 
 #endif /* SensorFusionAlgorithm_h */
